requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
flask==3.0.3
confluence-rest-library==1.2.2
pytest==8.1.1
black==24.4.0
flake8==7.0.0
pre-commit==3.7.0
httpx==0.23.3
python-dotenv==1.0.1
pytest-asyncio==0.23.6
pytest-mock==3.14.0
